sua lai storage link thành public link

+++ app/Repositories/ChildrenRepository.php
@@ -17,9 +17,13 @@ class ChildrenRepository
         $post = new $this->model;
         $post->content = $request['content'];
         if (isset($request['cover_img'])) {
-            $file = $request['cover_img'];
-            $path = $file->store('public/children');
-            $post->cover_img = Storage::url($path);
+            // Storage link not working in heroku
+//            $file = $request['cover_img'];
+//            $path = $file->store('public/children');
+//            $post->cover_img = Storage::url($path);
+            // Save images to /public/uploads
+
+            $post->cover_img = ImageHelper::store($request->file('cover_img'));
         }
 
         $post->title = $request['title'];
@@ -32,13 +36,19 @@ class ChildrenRepository
     {
         $post = $this->model::where('slug', $slug)->first();
         if ($post == NULL) abort(404);
-        if (isset($request['cover_img'])) {
+        if (isset($request['cover_img']) && $request->hasFile('cover_img')) {
+//            $fileExtension = $request->file('cover_img')->getClientOriginalExtension();
+//            $fileName = time() . "_" . rand(0, 9999999) . "_" . md5(rand(0, 9999999)) . "." . $fileExtension;
+//            $uploadPath = public_path('/uploads');
+//            $request->file('cover_img')->move($uploadPath, $fileName);
+            $post->cover_img = ImageHelper::store($request->file('cover_img'));
             if (file_exists(public_path($post->cover_img)) && $post->cover_img != $this->default_cover_img) {
                 unlink(public_path($post->cover_img));
             }
-            $file = $request['cover_img'];
-            $path = $file->store('public/children');
-            $post->cover_img = Storage::url($path);
+
+//            $file = $request['cover_img'];
+//            $path = $file->store('public/children');
+//            $post->cover_img = Storage::url($path);
         }
         $post->content = $request['content'];
         $post->title = $request['title'];
